+++ CHANGELOG.rst
@@ -3,7 +3,9 @@ Release Notes
 
 Forthcoming
 -----------
-* ...
+* [composites] sequences w/o memory, `#325 <https://github.com/splintered-reality/py_trees/pull/325>`_
+* [composites] selectors with memory, `#324 <https://github.com/splintered-reality/py_trees/pull/324>`_
+* [display] unicode trees are unicode, not ascii `#324 <https://github.com/splintered-reality/py_trees/pull/324>`_
 
 2.1.4 (2021-02-24)
 ------------------